Fix reviewer comment
